chore(cypress): extend test suite to cover usage of live quiz templates in catalog collections (#4557)
